astraweb/requirements.txt
Django>=5.0.0
gunicorn
psycopg2-binary
redis
celery
django-celery-beat
pyotp
qrcode
Pillow
django-user-agents
ua-parser
bcrypt
Django>=4.2
djangorestframework
django-axes
django-ratelimit
django-otp
django-otp-totp
qrcode
redis
django-redis
python-dotenv
